# Significantly improve initial load perf with "hybrid" execution + lazy compilation