--- a/src/engine/test/wish_engine/scripting_api_test.cljs
+++ b/src/engine/test/wish_engine/scripting_api_test.cljs
@@ -1,6 +1,6 @@
 (ns wish-engine.scripting-api-test
   (:require [cljs.test :refer-macros [deftest testing is]]
-            [wish-engine.test-util :refer [eval-form eval-state]]
+            [wish-engine.test-util :refer [eval-form eval-state eval-state-seq]]
             [wish-engine.core :as core]
             [wish-engine.scripting-api :as api]))
 
@@ -188,27 +188,30 @@
 
   (testing "Support using items-from-list for feature options"
     (let [{{f :crew-member} :classes :as state}
-          (eval-state
-            '(do
-               (declare-features
-                 {:id :weapon
-                  :values (items-from-list :weapons)})
+          (eval-state-seq
+            '[(declare-features
+                {:id :weapon
+                 :values (items-from-list :weapons)
+                 :! (on-state
+                      (provide-features
+                        {:id :weapon/spare
+                         :values (items-from-list :weapons)}))})
 
-               (declare-list
-                 :weapons
-                 {:id :knife
-                  :! (on-state (provide-attr :knife true))}
-                 {:id :pistol
-                  :! (on-state (provide-attr :pistol true))}
-                 {:id :rifle
-                  :! (on-state (provide-attr :rifle true))}
-                 {:id :vera
-                  :! (on-state (provide-attr :vera true))})
+              (declare-list
+                :weapons
+                {:id :knife
+                 :! (on-state (provide-attr :knife true))}
+                {:id :pistol
+                 :! (on-state (provide-attr :pistol true))}
+                {:id :rifle
+                 :! (on-state (provide-attr :rifle true))}
+                {:id :vera
+                 :! (on-state (provide-attr :vera true))})
 
-               (declare-class
-                 {:id :crew-member
-                  :! (on-state
-                       (provide-features :weapon))})))
+              (declare-class
+                {:id :crew-member
+                 :! (on-state
+                      (provide-features :weapon))})])
           state! (:! f)]
       (is (ifn? state!))
       (is (empty? (:attrs (state! {:wish-engine/state state}))))
@@ -217,6 +220,11 @@
                               {:weapon [:pistol]}
                               :wish-engine/state state})]
         (is (= {:pistol true}
+               (:attrs inflated))))
+      (let [inflated (state! {:wish-engine/options
+                              {:weapon/spare [:vera]}
+                              :wish-engine/state state})]
+        (is (= {:vera true}
                (:attrs inflated)))))))