Previously, loading the SRD took ~1800-2000ms, compared to ~550-600ms with the old implementation. Yikes.

First, instead of compiling *everything* into javascript, we can shave off ~30% (down from 2000ms to 1200ms) by only actually compiling things that need to be executed, for example `(let)` and `(for)` forms, etc., and then passing the results directly into the appropriate function. Static maps don't need to be compiled, and in fact even maps that have a `:!` key don't themselves need to be compiled, just the value of their `:!` key.

Next, since we don't actually *need* most of the fn values, we can lazily compile them all, wrapping the value in a fn that performs the compilation on demand. This shaves off another 40% (from 1200ms to ~700ms) and puts us in the ballpark of the original method, but keeping all the advantages of this one. 

Hurray!